-Modify INT to INTEGER

+++ src/jsMain/kotlin/DisplayMode.kt
@@ -1,6 +1,6 @@
 enum class DisplayMode(val chunkSize : Int) {
     BINARY(8),
-    INT(3),
+    INTEGER(3),
     HEX(2),
     ASCII(1)
 }
